Move `BarAndLabel` into `web-common/components`
