+++ web-common/src/components/BarAndLabel.svelte
